feat: move utilites methods to a single class
