+++ port/src/main/java/com/happening/kafka/Utils.java
@@ -1,11 +1,19 @@
-package com.happening.kafka.utils;
+package com.happening.kafka;
 
 import java.util.Collections;
 import java.util.Map;
 import java.util.Properties;
 
-public final class PropertiesUtils {
-    private PropertiesUtils() {
+public final class Utils {
+    private Utils() {
+    }
+
+    public static String getErrorMessage(Throwable t) {
+        var cause = t.getCause();
+        if (cause != null && cause.getMessage() != null) {
+            return cause.getMessage();
+        }
+        return t.getMessage();
     }
 
     public static Properties toProperties(Map<Object, Object> map) {
@@ -15,4 +23,5 @@ public final class PropertiesUtils {
         result.putAll(map);
         return result;
     }
+
 }
