--- port/src/main/java/com/happening/kafka/utils/ErrorUtils.java
@@ -1,16 +1,0 @@
-package com.happening.kafka.utils;
-
-public final class ErrorUtils {
-
-    private ErrorUtils() {
-    }
-
-    public static String getMessage(Throwable t) {
-        var cause = t.getCause();
-        if (cause != null && cause.getMessage() != null) {
-            return cause.getMessage();
-        }
-        return t.getMessage();
-    }
-
-}
